added InvestmentPlanController and more unit tests

+++ core/src/main/java/io/eternalwind/mira/core/models/InvestmentPlan.java
@@ -1,6 +1,5 @@
 package io.eternalwind.mira.core.models;
 
-import java.math.BigDecimal;
 import java.util.List;
 import org.springframework.data.mongodb.core.mapping.Document;
 
@@ -15,7 +14,7 @@ import lombok.experimental.SuperBuilder;
 @ToString(callSuper = true)
 @SuperBuilder(toBuilder = true)
 public class InvestmentPlan extends Entity {
-    public static record SectorWeight(String sector, BigDecimal weight) {}
+    public static record SectorWeight(String sector, double weight) {}
 
     private String name;
     private List<SectorWeight> sectorWeights;
